Cosmetic Changes
Changed GitHub color from the blue they use on the site to the purple
they use in the desktop app. I like it so much better, and it breaks up
monotony in the footer social menu. Hid h1's with class 'page-title'.
Changed the hover effect on portfolio thumbnails to include a
transparent-to-white gradient. Lowered the post meta to fit in the lower
40% of the thumbnail. Will need to write responsive CSS to adjust this
for smaller screens.